feat(kubernetes): make --prune respect --target
